Style fixes, script refactor, particles on admin
