+++ assets/js/website/navbar.js
@@ -1,3 +1,5 @@
+import {handleDropdown} from '../dropdown';
+
 document.addEventListener('DOMContentLoaded', () => {
     const cardElements = document.querySelectorAll('.hover-element');
     const navItems = document.querySelectorAll('.nav-item');
@@ -21,18 +23,6 @@ document.addEventListener('DOMContentLoaded', () => {
             cardElements[index].classList.remove('hovered');
         });
     });
+
     handleDropdown();
 });
-
-function handleDropdown() {
-    const navlangElement = document.querySelector('.nav-lang');
-    const dropdownElement = document.querySelector('.home-dropdown-menu');
-
-    if (!dropdownElement) {
-        throw new Error('46b745a7-c816-49a2-aaf3-50266c8cc366');
-    }
-
-    navlangElement.addEventListener('click', () => {
-        dropdownElement.classList.toggle('d-block');
-    });
-}
