+++ templates/admin/partials/topbar.html.twig
@@ -11,16 +11,28 @@
     </div>
     <div class="nav-lang nav-lang-admin">
         {% if app.locale != 'pl' %}
-            <a class="nav-lang-link" href="{{ path('app_language_admin_changer', { '_locale':'pl' }) }}">
-                <i class="fi fi-pl fis fi-lang"></i>
-                <div class="tooltip">Polski</div>
-            </a>
+            <span class="nav-lang-link">
+                <i class="fi fi-gb fis fi-lang"></i>
+            </span>
         {% endif %}
         {% if app.locale != 'en' %}
-            <a class="nav-lang-link" href="{{ path('app_language_admin_changer', { '_locale':'en' }) }}">
-                <i class="fi fi-gb fis fi-lang"></i>
-                <div class="tooltip">English</div>
-            </a>
+            <span class="nav-lang-link">
+                <i class="fi fi-pl fis fi-lang"></i>
+            </span>
         {% endif %}
+        <ul class="home-dropdown-menu">
+            <li>
+                <a class="nav-lang-link" href="{{ path('app_language_admin_changer', { '_locale':'pl' }) }}">
+                    <i class="fi fi-pl fis fi-lang"></i>
+                    Polski
+                </a>
+            </li>
+            <li>
+                <a class="nav-lang-link" href="{{ path('app_language_admin_changer', { '_locale':'en' }) }}">
+                    <i class="fi fi-gb fis fi-lang"></i>
+                    English
+                </a>
+            </li>
+        </ul>
     </div>
 </div>
