+++ templates/admin/security/login.html.twig
@@ -8,6 +8,7 @@
     {% endblock %}
 </head>
 <body class="h-100">
+    <div id="particles-js"></div>
     <div class="login-panel">
         <form method="post">
             {% if error %}
@@ -41,5 +42,8 @@
 </html>
 
 {% block javascripts %}
+
+    <script src="http://cdn.jsdelivr.net/particles.js/2.0.0/particles.min.js"></script>
+    {{ encore_entry_script_tags('particles') }}
     {{ encore_entry_script_tags('app') }}
 {% endblock %}
