+++ templates/website/home.html.twig
@@ -17,5 +17,5 @@
     {{ parent() }}
 
     <script src="http://cdn.jsdelivr.net/particles.js/2.0.0/particles.min.js"></script>
-    {{ encore_entry_script_tags('website/particles') }}
+    {{ encore_entry_script_tags('particles') }}
 {% endblock %}
